[iter14] separate storage config from server config
