+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/ospiem/mcollector
 
-go 1.21.5
+go 1.21.3
 
 require (
 	github.com/caarlos0/env/v9 v9.0.0
