+++ internal/storage/config/config.go
@@ -1,0 +1,10 @@
+package config
+
+import "time"
+
+type Config struct {
+	FileStoragePath string `env:"FILE_STORAGE_PATH"`
+	DatabaseDsn     string `env:"DATABASE_DSN"`
+	Restore         bool   `env:"RESTORE"`
+	StoreInterval   time.Duration
+}
